Fix up/down fields in FV::Div_par_K_Grad_par
Previously, fup/fdown and Kup/Kdown fields were defined but not used.
The inputs fin and Kin were copied to f and K before checking
hasYupYdown(), but since yup_field and ydown_field are not copied in the
Field3D assignment operator, the yup/ydown branch could never be used.