Make BoundaryOp and BoundaryRegion members const, deprecate BoundaryWidth
Make several member variables const: apply_to_ddt of BoundaryOp and
width and label of BoundaryRegion.

The (probably never used) functionality of the BoundaryWidth modifier
has been replaced with 'width' keyword to boundary conditions. The
BoundaryWidth modifier needs to keep changing BoundaryRegion::width,
which we are making const, so can no longer be used. Trying to use a
BoundaryWidth modifier now throws an exception with an instruction to
use the 'width' keyword instead.