Revert "Simplify the choice of how many guards to use."
This reverts commit 4c5d32b0c9ccb858bab576e4c47c5d0c95a3392b.

This is currently necessary for gcc compilers using oldish libstdc++
This is evident on travis which uses the 4.8 contemporary version (for
both v4.8 and v7 of the compiler). This shouldn't be necessary for newer
versions and has been tested as working with gcc 5.4 using v6 of the
libstdc++

Hopefully we can reintroduce the commit that this reverts in the future
as we change the versions of things that we support.