Support BoutOutputs caching option in boutcore *.fromBoutOutputs
Previously the 'tind' option was passed to the BoutOutputs object, but
if 'caching' is set to true, then the data from the first time a field
is collected is stored, so the data from the first 'tind' passed would
be returned each time.

Now handle the 'caching=True' and 'caching=False' cases separately: for
True collect full arrays and return the slice at tind; for False just
collect the data at tind.