Fix implementation of 'width' in boundary conditions, use BoundaryWidth
Cannot implement 'width' properly with member of BoundaryOp.

Un-deprecates BoundaryWidth modifier, since this is now implemented (and
needed to implement the 'width' keyword).

Add 'copy' method to BoundaryRegion. Allows creating a new
BoundaryRegion with a different width. Default value for 'wid' argument
is -1, in which case xstart/ystart is used and the returned object is an
exact copy of the original.

BoundaryWidth modifier now has a member 'unique_ptr<BoundaryRegion>
bndry' which is copied from the input BoundaryRegion, but with the width
changed.