Clear existing boundaries in FieldData::setBoundary()
Otherwise if setBoundary() is called twice, the first set of boundary
conditions will be applied and then the values overwritten by the second
set. This is inefficient, and might cause unexpected behaviour, e.g. if
the second set of boundary conditions specified "none" for some
boundaries where the first set did not.