Allow initialization of Coordinates with Mesh::addCoordinates(location)
Instead of creating Coordinates objects when Mesh::getCoordinates is
called, create them during initialization. As a temporary workaround to
maintain backward compatibility, initialize all locations in
BoutMesh::load() so that they are always available.

Allow the user to explicitly request a location to be available, by
calling Mesh::addCoordinates(location) for each location required. This
is optional in v4.2 but will be required from v4.3.

When Coordinates::geometry() is called, assume that staggered location
Coordinates objects need to be recalculated because the user has changed
the CELL_CENTRE version.

As a temporary workaround, replace geometry() with
geometryNoRecalculate() and geometry() now calls geometryNoRecalculate
and then recalculates the staggered versions. This means geometry() can
be reset back to not recalculating in next (which is OK because we will
add the requirement to call addCoordinates() after calling geometry())
without changing the function signature.