3D test for LaplacePetsc3DAmg
Basic so far, only in slab with orthogonal coordinates, but hacked
g22=g_22=0.9 so some y-derivative terms are included (hack because the
covariant and contravariant metrics are not inverses of each other like
this, but this does not matter for the test). So far also periodic in
y-direction (no 'open-field-line' boundaries).