Hypnotoad: fix unstable integration method used to calculate zShift
IDL's INT_TABULATED method uses a 5-point Newton-Cotes integration
formula. This multi-point integration introduces oscillations in the
integral when integrating across domain boundaries (e.g. pfr->core)
where there are discontinuities in the integrand or very near X-points
where the integrand is almost singular.

Add the option to replace INT_TABULATED with a simple trapezoid
integration, which is stable, in the function 'inty'. Use this option
('simple') when inty is called to calculate zShift.