Use localmesh for d2x and d2y, and interpolate them to location
Previously the localmesh was not passed through to the constructor of
d2x and d2y in Coordinates::geometry(), so they used the global 'mesh'.

Also, when d2x/d2y are read from the mesh, they are at CELL_CENTRE, so
we need to interpolate them to the location of the Coordinates.