Remove uses of Coordinates objects from ParallelTransform constructors
dy is needed by the FCIMap constructor and dz by the ShiftedMetric
constructor. Read directly from the Mesh data source instead of reading
from Coordinates objects. This avoids the need to create the Coordinates
objects until the ParallelTransforms have finished being created, so
coordinate_system can be set for the member fields of Coordinates.

To reduce code duplications, adds a 'getnz()' method to Mesh, which
handles the 3 possible ways of specifying dz in input.