Important fix for proportional N, T profiles
When input equilibria do not specify separate density and temperature
profiles, there are several options for how these should be set. One
of these options (3 in hypnotoad) is to keep the same shape for both
profiles. This code had a bug, which resulted in the output pressure
profiles being proportional to the square of the input pressure
profiles.

This bug affected both hypnotoad and the pdb2bout.pro processing used
for input from DSKGATO or ELITE inputs.

Thanks to Seto-san for finding.