Add a second template argument to Array used to specify what type to use as the memory backing/container
Defaults to ArrayData so no change required to user code. Allows us to
remove the need to pick the backend at configure time (BOUT_FLAGS time)
and allows us to mix and match (to some extent).

Note this may make it harder for us to cleanup the array stores as we'll
have a different store for every {data type, backend} combination used,
which we currently have no easy way to discover.