FieldFactory: Separate parsing, looping stages in create{2,3}D
To create a field in the RHS function, for example sources as a
function of time, currently we have to do something like:

```
FieldFactory::get()->create3D( Options::root()["vorticity"]["source"],
                               &Options::root()["vorticity"],
                               mesh,
                               CELL_CENTRE,
                               time );
```

which involves looking up an options string, parsing it and then
evaluating the resulting generator over the mesh. This isn't as bad as
it looks, because the string is hashed in the parser so is in a lookup
table.

To clean this up a little, this splits `create3D` into the parse step,
and the part which calls the generator. This adds additional functions
to the API which take a `FieldGeneratorPtr` rather than a
`std::string`.

The above example becomes:

```
FieldGeneratorPtr generator =
       FieldFactory::get()->parse(Options::root()["vorticity"]["source"],
                                  &Options::root()["vorticity"]);

...

FieldFactory::get()->create3D( generator,
                               mesh,
                               CELL_CENTRE,
                               time );
```

Adds some documentation but no tests yet.