Bug fix: remove communications from Coordinates ctor
Fixes #1369

`geometry()` is no longer called from the `Coordinates`
constructor. This means that the Christoffel symbols and non-uniform
corrections will not be available until either `geometry()` or the new
`calcChristoffelSymbols()`/`calcNonUniformCorrections()` are called.

These functions are called on demand in the library functions where
the relevant variables are needed.

Note that this is technically a breaking change, but user code is
unlikely to be affected. If it is, the fix is a one liner:

    mesh->getCoordinates()->geometry();