Recalculate staggered Coordinates when geometry() is called
Coordinates at staggered locations are calculated by interpolating the
CELL_CENTRE Coordinates. If the CELL_CENTRE Coordinates are updated (in
which case geometry() must be called), then the staggered Coordinates
need to be re-calculated. This commit adds a method to Mesh to
re-calculate any staggered Coordinates that have been added to
Mesh::coords_map, and calls the method from Coordinates::geometry() when
location==CELL_CENTRE. The location test is necessary because geometry()
also needs to be called when constructing the staggered Coordinates.