Honor region if possible
Not honoring the region breaks cross-interpolation.
This requires that the user has set the corner cells.
With that, we still do not interpolate into the boundaries in the
direction we are interpolating, but that is another issue.
At least if interpolating in Z direction, not honoring the boundaries is
wrong, as it breaks the double interpolation, which is implemented.