Make application of INVERT_SET at outer boundary clearer in cyclic solver
The condition at the outer boundary was
(xe - ix < outbndry) && (outer_boundary_flags & INVERT_SET)
&& mesh->lastX())

This form is misleading, as the inner boundary condition does not depend
on 'xs'. This commit changes the expression to:
(mesh->LocalNx - ix - 1 < outbndry)
&& (outer_boundary_flags & INVERT_SET) && mesh->lastX())

Both are correct, as xe=mesh->LocalNx-1. However, it is possible this
will change in future because solving in more than one guard cell is not
necessary, and if the value of xe was changed (e.g. to mesh->xend+1)
then the condition on the outer boundary would not be correct.