Hypnotoad follow separatrix further from X-point
When following lines along separatrix away from the X-point, continue
until both R and Z have reached an extremum. This should take the line
at least a quarter and at most half way around a closed contour. The
previous behaviour risked stopping after a very short distance if the
X-point was not at the top or bottom of the contour.