Hypnotoad: use INTERPOLATE instead of fft_interp on contours
Using INTERPOLATE is consistent with how contours are treated elsewhere,
which can prevent errors (for example the projected poloidal location of
the X-point being slightly different when calculated when fft_interp or
INTERPOLATE is used to find the starting position, leading to a position
being found on the wrong side of the X-point).

Also, fft_interp is likely to be inaccurate on contours, as their points
can be very unevenly spaced. INTERPOLATE uses linear interpolation which
is more robust.