Hypnotoad: more flexible transitions in nonorthogonal grid generation
Previously, the nonorthogonal grid generator changed the radial grid
direction from the boundary vector at one end of a region, to the normal
to flux surfaces half way through the region, and then to the other
boundary vector at the far end of the region. This could cause issues,
for example if the line normal to the flux surfaces half way through the
region intersects with one of the region boundaries. Instead, use both
vectors with separate weights. The weights are now set to decay with the
power 2.7 instead of 1.35, because the parameter under the power now
increases/decreases half as quickly (but extends the full length of the
region, not just half).