Make FieldFactory and ExpressionParser const correct
FieldFactory::options is a hidden argument to FieldFactory::resolve
and can be overridden during a call to parse/create2D/create3D, which
is why it needs to be marked `mutable`. Note that it is a mutable
pointer to a const Options -- it's the pointer that's mutable, and the
object that is const!