Hypnotoad: add setting to change power of nonorthogonality decay
The radial direction of the nonorthogonal grid is calculated by taking a
weighted average of the two vectors defining the directions of the ends
of the grid section and the vector orthogonal to the flux surface. The
weight of each end vector reduces as a power of the fraction of the
poloidal index-space distance remaining in the grid. The new setting
allows this power to be changed, so that the grid approaches orthogonal
faster or slower.