Add test for non-uniform boundaries
The test seems to be broken, as the convergence is not what it is
supposed to be.
This is most likely a bug in getting the correct dy and realy, which
seems to limit the accuracy.
with the realy() implementation, the test converged to more reasonable
results, i.e. an o4 was never limited to first order convergence.