Extrapolate metric components into boundary/guard cells where necessary
Metric components must be extrapolated into boundary cells to make sure
that their gradients are continuous and smooth at the boundary.

Also, when using ParallelTransformIdentity, the integrated shear I is
discontinuous at the poloidal boundary of the core, closed-field-line
region, so metric components must also be extrapolated at this boundary.