Fix extrapolation of staggered Coordinates quantities
There is a fix in interpolateAndExtrapolate to interpolate the xend+1
(at CELL_XLOW) or yend+1 (at CELL_YLOW) boundary point. However, this
causes a bug when interpolateAndExtrapolate is used just to extrapolate
the boundary cells, as it is for J and Bxy. In this case (if the
location of the field is the same as the location argument) the extra
interpolation must be skipped, leaving just the extrapolation (note the
extrapolation does and should still skip the xend+1 or yend+1 point
where interpolation would have been done, because the field, e.g. J,
given as input has been calculated correctly from metric components that
were interpolated to that position).