Move ParallelTransform ownership from Mesh to Coordinates
This means the ParallelTransform objects naturally have a specific
location, defined by their owning Coordinates object.

Also, require zShift and zlength to be passed to the ShiftedMetric
constructor, so Coordinates is responsible for getting zShift from the
grid (at the correct location). Passing zlength means that the
ShiftedMetric constructor does not require a Coordinates object, and
therefore can be created in the Coordinates constructor.

Need to pass an Options* pointer to the Coordinates constructor from
the owning Mesh object, so Coordinates can use it to initialize the
ParallelTransform.