Correct zShift in guard cells so ShiftedMetric doesn't need TwistShift
To make zShift continuous into the guard cells where there is a branch
cut in the poloidal domain from theta=0->theta=2pi, we need to
add/subtract ShiftAngle (which is the total zShift for a 2pi poloidal
circuit).

Add methods to Mesh/BoutMesh to indicate where there is a branch cut and
get the value of ShiftAngle.