Hypnotoad: better refinement of starting locations in nonorthogonal
Previously, for nonorthogonal grids the starting positions (points along
the separatrix) were refined only when the grid was being created in the
core. This commit moves the refinement to a position (previously
commented-out) in grid_region_nonorth which ensures it is done in every
case. This makes non-orthogonal grid generation more robust, especially
around x-points.