Change variable names for output from staggered Coordinates
Previously, the variables in Coordinates would be output with names that
did not change if the Coordinates object was at a staggered location.
This could have resulted in variables being over-written if
Coordinates::outputVars was called on two different Coordinates objects.

This commit adds a suffix "_"+toString(location) if location is not
CELL_CENTRE.