Fix string comparison in ParallelTransform::checkInput
Apparently strings returned from Mesh::get have a length 1 greater than
if you just type in a string literal, so the operator== comparison was
failing. Converting to c_str and using strcmp gets around this (as
strcmp just compares up to the first null byte, so does not see the
different lengths).