Revert "Use non-const BoutReal arguments in gen_fieldops.py/gen_fieldops.jinja"
This reverts commit 80525bab6332e2724dc2310241ba8e7fdcad053a.

It does make sense to have 'const' qualifier for pass-by-value arguments
only in the implementation and not in the interface. The const does not
affect the interface, so does not need to be there, but affects the
local variable inside the scope of the implementation. See #1655 for
explanation from @dschwoerer.