Counters for timers, makes them correct with multiple Timer objects
Previously, if multiple Timer objects were created with the same label,
the timings would not be correct. The timer should run as long as any of
the objects are in scope. This commit makes this happen by adding a
counter to the timer_info struct, and only stopping the timer when the
counter is 0.