Change Options::at interface
The optional input argument is now a value which the result should be
similar to. This is so that `Field` types can be requested, and the
metadata like `Mesh` pointer and cell location will be taken from the
input.

No (non-test) code uses this interface yet, so should be ok to change.