Fix unit tests
In OptionsTest, use FakeMeshFixture so we can use the
bout::globals::mesh which has Coordiates objects created. Avoids
implicit creation of Coordinates when getCoordinates is called, which
lead previously to a segfault because the locally FakeMesh doesn't have
a GridDataSource. We can't resolve this by calling
'fieldmesh->setCoordinates(nullptr)' because create3D requires a
ParallelTransform, which belongs to the Coordinates.