Multigrid: comment out a warning, replace some checks with ASSERT1
Comment out a conditional that prints a warning for a case that should
already be handled (the value of 'a1' is too small, which should only
happen if the rhs is nearly zero everywhere). [There's another check for
a new 'a1' a bit below, I'm not sure this is safe to remove].

Replace a few other checks with ASSERT1 so we don't keep them in fully
optimized configuration. The check for positive 'ini_e' should be safe
because it's the result of a dot product which should always be positive
unless there's a bug in the vectorProd() method. I'm not sure the other
twor are always safe, but the checks can be re-enabled by configuring
with CHECK>=1.