Simplify CMake SUNDIALS config by directly checking integer type
Adds `bout::utils::function_traits` to allow getting types of function
arguments. This is then used to get the type of the argument for the
various SUNDIALS user functions. In later versions of SUNDIALS this is
provided by the typedef `sunindextype` which we can use directly.

This removes the need for the preprocessor macros CVODEINT etc