Add FieldPerp support to collect.py
Note: read FieldPerp only from files with yindex_global>=0 - A negative
yindex_global indicates an invalid FieldPerp in that file.  There should
be only one pe_yind where the FieldPerp was written, which will have
yindex_global>=0, and these should be the only ones read.

Previously attributes were always loaded from processor 0, but this is
wrong for FieldPerps that are not not valid on that processor. For
FieldPerps, check the yindex_global, and get attributes from the first
processor where it is valid (i.e. part of the grid, including guard
cells, on that processor). Also check that every processor where it is
valid has the same y-processor-index.