More FieldPerp support in datafile.py
Tries to ensure that the 'bout_type' attribute gets read/set correctly.

When the input to DataFile.write is a BoutArray, we can get the variable
type, and therefore dimensions, from the 'bout_type' attribute. This
means we don't have to rely on just counting the number of dimensions,
which cannot distinguish between Field2D and FieldPerp. Enables writing
of FieldPerp with the correct {'x','z'} dimensions.

Fix check for string attribute in DataFile_HDF5