Make restart.redistribute work for FieldPerps
Notes:
- Update MXSUB, MYSUB in restart.redistribute. Previously kept old
values of MXSUB and MYSUB, this caused problems in using 'collect' to
get the data, instead of the previous reading direct from 'DataFile's.
Using collect is simpler though, and should be more robust. The new
approach means MXSUB and MYSUB are consistent with the processor layout
of the restart files.
- When writing a null FieldPerp (because the processor being written for
does not contain the yindex_global of the FieldPerp), make sure
mesh->YLOCAL(yindex_global) is always negative so that the FieldPerp is
always recognized as invalid.