Do not calculate parallel slices for field-aligned fields
It is incorrect to calculate yup/ydown fields for an input that is
already field-aligned. In calcParallelSlices, if
'f.getDirectionY() == YDirectionType::Aligned', then return without
calculating parallel slices. Do not throw an exception, because: a
simulation might mix field-aligned fields with fields using parallel
slices; calcParallelSlices() may be called in communicate(); and a
field-aligned field might need to be communicated.

In the y-derivatives, before calling a YOrthogonal method, add ASSERT1
checks that the input or inputs have YDirectionType::Standard.