Tidy up direction-type setters
setDirectionType(d) methods of Field3D, Field2D and FieldPerp that
return the field (allowing method chaining) are the interface to set the
YDirectionType. Add Field2D::setDirectionType(d) in case it is wanted in
a template function which needs to call it on Field3D/FieldPerp. Field2D
version ASSERT1's that the argument is YDirectionType::Standard, which
is the only valid value for Field2D as field-aligned and orthogonal
Field2Ds would be identical.

Remove the setDirectionTypeY and setDirectionTypeZ methods of Field.
Functionality of setDirectionTypeY is provided by the derived-class
methods. setDirectionTypeZ is not needed anywhere at the moment, and
should be provided by similar derived-class methods if/when it is
needed. Make Field::directions protected instead of private so that the
derived classes don't need the setters.