Add RGN_NOCORNERS, RGN_XGUARDS, RGN_YGUARDS and RGN_ZGUARDS by default
Add RGN_NOCORNERS which excludes the corner guard cells. This is useful
for checks which want to allow the corner guard cells to be set to
BoutNaN. Also add regions containing just x- y- or z-guard cells, which
are useful in constructing RGN_NOCORNERS.

Add RGN_NOZ for Field2D. Consistent with having a RGN_NOY for FieldPerp.
May be useful in template functions which want to get RGN_NOZ from any
type of Field.

Rename what was RGN_NOCORNERS in invertable_operator.hxx to
RGN_WITHBOUNDARIES as it has a different purpose - it excludes internal
guard cells, but includes some boundary cells.