Fix interpolation of staggered zShift
zShift needs to have its guard cells correctly initialized. Previously
when initializing a staggered zShift, if the staggered zShift was not
present in the input grid then an unstaggered zShift was read from the
grid and interpolated to the staggered location; it had not had its
guard cells corrected and this led to errors. This commit applies the
guard cell fixes to the cell-centre zShift before interpolating to
staggered grid.