Fix GridFromOptions setting default value in Options
The default argument to `Mesh::get` is new. Existing calls may be
using the return value to set a default, which may or may not be
recorded in the `Options` tree. We don't want the default value passed
to `GridFromOptions::get` to get recorded in case we don't end up
using it. This is for the benefit of the `BOUT.settings` file that
records the settings actually used during a simulation