Enable multi-line expressions in input options
Long expressions can become hard to read in the options file.
This adopts the Python approach, where if there are unbalanced
brackets ('()' or '[]') then the expression continues on the next
line.

In the BOUT.inp file we can therefore write

```
value = (something
         + more
         - things)
```

or define complex expression using a Context scope:
```
result = [a = expression,
          b = something](
            {a} + {b})
```