Recursion, Turing completeness
By default recursive expressions are not allowed in the input options.
What is committed here adds an input option to allow recursion:
```
[input]
max_recursion_depth = 10  # 0 = none, -1 = unlimited
```
By putting a limit on the depth we avoid the halting problem at least,
and expressions should (eventually) terminate, preferably before they
overflow the stack.

As is traditional, I present to you the Fibonacci sequence:
```
fib = where({n} - 2.5,
            [n={n}-1](fib) + [n={n}-2](fib),
            1)
```
so if `n` = 1 or 2 then `fib` = 1, but if `n` = 3 or above then the
recursion fun begins.

And another ill-considered language is born.