Fix writing of BoutArrays in DataFile_netCDF Python class
Previously tried to check if the input variable was a BoutArray with
't=="BoutArray"', but 't' has already been changed to the data-type of
the array entries so the condition always evaluated to 'False'. Instead
just try to use 'data.attributes['bout_type']' and catch the
'AttributeError' that is thrown if the variable is not a BoutArray (or
'bout_type' has not been set).