Construct Option with initializer_list
This enables initializer lists to be used to create custom options,
to pass to solvers or other components. This is a slightly more
compact way to create options trees than the existing method.

Unfortunately requires non-template constructors for the accepted
types, but these are quite simple.

Includes tests and documentation.