Finite difference implementation of LaplaceXY
The existing implementation of LaplaceXY uses a finite-volume
discretisation. This commit introduces an implementation using the same
finite-difference discretisation as the 'Laplace(Field2D)' and
'Grad(Field2D)*Grad(Field2D)' operators would use.

Also, when using the finite-difference implementation, values for the
boundary conditions are passed in the same way as for the 'Laplacian'
solvers, i.e. as the value in the boundary cell, instead of the mean of
the boundary and neighbouring grid cells.