Wrote interface and tests for Laplace 3D AMG solver
I have added an object-oriented interface between BOUT and PETSc which
will allow for easier conversion between Bout Fields and PETSc vectors
and matrices. In particular, the interface uses BOUT indexing and then
internally handles the conversion to PETSc global indexing. I have
written unit tests for this interface to ensure it behaves as
expected. In particular, I have checked that it can handle indexing
properly for parallel simulations.

As all tests are run in serial, this required quite a lot of work. I
created a new fake mesh class which acts as though it is on just one
processor of many. By creating a group of these meshes they can pass
information to each other, emulating communication between
processors. I also needed to fake MPI calls, which required creating
virtual methods for them on the mesh class. These are then overridden
in the fake mesh implementation. I also structured the classes in my
PETSc interface so that fake classes could override certain methods
where it is necessary to make the pretend-parallelism work.

This PETSc interface will make implementation of the Laplace solver
much simpler. That solver follows the interface of existing ones,
except that it now accepts options for setting the boundary conditions
in the y direction. I also added some methods to get the matrices
being solved. This is to aid in unit testing; to ensure matrices are
correct, I compare the results of matrix-multiplication with those of
applying the desired Laplace operator in the foreward direction. This
is done both for the default coefficient values in the operator and
after setting each of them to a custom value. Finally, the solver as a
whole is tested to ensure it can properly invert the foreward
operator. Note that this is all for simple geometries. More complex
ones will be tested in the integration tests.