Require inputs to to/fromFieldAligned to be in the correct space
That is, to/fromFieldAligned are never no-ops.

Also requires the FieldPerp overloads to be over RGN_NOX: this region
is essentially hardcoded currently, so accepting other regions could
be confusing.

The y-derivatives still accept fields in either space, and will
transform unaligned fields if necessary, returning the result in the
same space as the input field. The flow derivatives now require both
fields to be in the same space.