Implemented working indexer class for use with PETSc interface.
This required me to make a number of changes to the fake parallel mesh
class I'd added for testing. I had made numerous errors in the
implementation of that and had overlooked a few problems. However, the
current state of the code will now compile and all indexer tests pass.