Flip signs of coefficients for free_o3 boundary condition in LaplaceXY
Solution should be identical since the rhs for the rows applying the
boundary condition is 0, but convergence seems to be much better with a
TORPEX X-point grid file this way, even though the test in
test-laplacexy was OK before. Possibly having a positive diagonal matrix
entry is a good thing?