Fix non-unique 'count' bug in PetscLib
Was using 'count' as a unique identifier, but 'count' is decremented
when a PetscLib is destroyed. This commit adds 'unique_id' which is
incremented each time a PetscLib is created, and never decremented.

Also some small tidy-ups: forward-declare Options; make sure KSP is
defined in the no-PETSc branch; add an empty body to
createKSPWithOptions in the no-PETSc branch.