Added additional optional arguments to Laplace routines
Changes made in next to how corner boundary conditions are handled in
D2DXDY broke the forward operator used in tests of the Laplace3DAmg
solver. It resulted in the x-boundaries being undefined prior to
calling DDX, causing NaNs to appear in the interior of the grid. I
have added some additional arguments which allow the calculation to be
performed in the old way if desired, but without changing default
behaviour from what exists on branch next.