Refactored MPI wrappers to be on own class rather than BoutMesh
This involved defining a class MpiWrapper and creating a global
instance of this. I changed all MPI calls to use the methods on this
class rather than the direct MPI calls. Files which include the PETSc
headers provded slightly difficult, as PETSc defines macros to overwrite MPI
calls in a way which is incompatible with my wrapper. These seem to
just be to provide diagnostic information and are not strictly
necessary, so my solution was to undef these macros where needed.
In addition to this refactoring, I changed the FakeParallelMesh class
which is used when testing the PETSc indexers to use a FakeMpiWrapper
class rather than override the MPI methods itself.