Ensure parallel_transform is only read from grid files, not from options
Previously 'mesh:parallel_transform' could be read from options in the
input file, causing unexpected/incorrect behaviour, as
'parallel_transform' was intended only for checking that a grid file is
compatible with the selected ParallelTransform. This commit adds a
method to Mesh to check whether the grid source is a file, and only gets
'parallel_transform' from the grid if it is, preventing
parallel_transform from ever being read from options.