Made some PETSc interface tests conditional on using debug version
Some of the PETSc tests deliberately call PETSc functions with invalid
input, to check that the correct error behaviour occurs. However, this
produces undefined results when linking against a version of PETSc
that was not compiled for debugging. As such, these tests are now
wrapped in conditional-compilation flags so they will only be included
when linking against a build of PETSc which does support debugging.