Modified PETSc interface to be thread-safe
The PETSc library is not thread-safe, meaning that calls to it must be
in critical blocks of code. Most of the PETSc calls in the interface
are unlikely to be performed from a parallel region, but those for
getting and setting vector and matrix elements very likely would be
called from a thread-parallel loop. These calls were thus given the
`critical` pragma.

When setting up the GlobalIndexer object, some loops must be performed
in serial as they rely on a counter variable.

One loop in the PETSc matrix tests required that a reduction be specified.