Test for 3d Laplace_perp inversion using a gridfile
Allows the 3d Laplace_perp solver to be tested in X-point geometries.

The test creates field-aligned input data by creating some initial data
at a few poloidal positions (outer/inner midplane and targets), then
applying parallel diffusion, then applying perpendicular diffusion.

A grid file is required for the test, but not provided in the repo: the
user must supply one for this test to run. The test is intended to be
compatible with any grid file.