Fix uses of GlobalNy and numberOfYBoundaries()
The correct value for the total number of y-boundary cells is
mesh.numberOfYBoundaries()*2*mesh.ystart
As there are two boundaries (upper and lower) containing mesh.ystart
cells each for each boundary.

Previously the factor of 2 was missed in several places in the recent
re-definition of GlobalNy. The use in Vol_Integral was missing a factor
of mesh.ystart as well.