Tidying up Petsc3dAmg to make more robust
A few small changes made while debugging integration tests. The
default tolerance for divergence was increased, as this was often
necessary for the solver to succeed. The block size of the matrix is
now explicitely set. I also realised that the way I was initialising
the field of zeros for the initial guess in the version of the solve
method where this is not provided by an argument would result in a
segfault. Therefore, I refactored that slightly.