Added new regions to mesh for use in PETSc interface
I've added additional boundary regions with the suffix _THIN which
differ from the exinst regions by

1. Including corner values (in y boundaries)
2. Only being one cell deep

In addition, the GlobalIndexer class creates a region from these and
RGN_NOBNDRY, called RGN_ALL_THIN. This allows far more compact
iteration over the grid cells to be stored in PETSc vectors. The new
regions will also be useful when setting boundary condition in various
solvers.

While doing this refactoring I caught an error in the routine on
BoutMesh which created the existing boundary regions. There was a call
to `firstX()` which should have been `lastX()` when determining
what indices to place in RGN_OUTER_X. I can only presume that this
hasn't been caught yet because no one has used these region when there
is more than one processor in the x-direction.

This refactor revealed that sometimes the GlobalIndexer class ended up
returning a stale instance between unit tests, due to a quirk of where
new values of `localmesh` were being allocated in memory. To avoid
this I have added an explicit `recreateGlobalInstance()` method, which
marks that the indexer corresponding to the global mesh should be
recreated the next time it is requested.