Add StandardFactory::ensureRegistered to defeat linker being clever
Not a particularly elegant solution, but ensureRegistered can be
defined by derived factories if the base class implementation file is
small/empty. Even having even an empty function in the .cxx file
causes the linker to keep all the symbols in that TU.

Calling ensureRegistered in the factory getInstance means
everything (should) get registered before the first time it is required.