Try to prevent race condition while building.
This improves the current situation, as before it could happen that
building failed, because the .a was not yet created. Now it waits for
it to be created. Switching to atomic mv rather then to create it
inplace, ensures that if the file is seen, it is valid.
This does not address the case when the .a is recreated, and the
dependend file is checked before whether the .a has changed.