Changes to ensure proper memory management in Laplace solvers
The new factory interface for Laplace solvers returns a unique_ptr
object rather than a raw pointer. Where necessary, I changed the integration tests to
be able to handle this.

I also discovered the way I was handling
finalising PETSc wasn't quite correct. I had an unnecessary instance
of `PetscLib` in the `GlobalIndexer` class. As the global indexer doesn't
get deallocated until the end of execution, this resulted in warnings
of "Instance of PetscLib still exist" when `BoutFinalise()` is
called. As such, I removed this member variable from that class.

With PETSc no longer prevented from finalising, I
saw that the PetscMatrix and PetscVector classes could end up trying
to deallocate PETSc objects after finalisation had already
occurred. This was fixed by changing the order in which member
variables of the class were declared, so that the destructor on
`PetscLib` would be called _after_ that on the `unique_ptr<Mat/Vec>`
object.

Finally, using `unique_ptr<Laplacian>` means that the destructor on
Laplacian operators is now always called before execution
terminates. However, depending on the scope of the solver, this might
not happen until after `BoutFinalise()` has been called, which would
force finalisation of PETSc and result in a segfault when the solver's
destructor tries to deallocate various PETSc objects. As such, it is
necessary to explicitly destroy the solver object before calling
`BoutFinalise()` in the integration tests that use a PETSc-based
Laplace solver.