Make end inclusive for 3-element list index argument to collect
When lists are passed to the `tind/xind/yind/zind` arguments of collect, for a 2-element list `[start, end]` the `end` index is interpreted as inclusive, as described in the docstring. 3-element lists `[start, end, step]` were not documented explicitly before, but should behave similarly. Previously 3-element lists treated `end` as exclusive.