Refactored PETSc interface to be more general.
Rather than assuming that a certain subsect of guard cells should be
included in a PETSc vector, it now computes which ones should be
included using an OperatorStencil. This required some fairly
significant refactoring so that the vector and matrix constructors now
must take an indexer object as an argument. The indexer stores the
information on which guard cells should be included and on the
sparsity pattern of a matrix.